+++ src/org/sosy_lab/java_smt/basicimpl/PackageSanityTest.java
@@ -21,6 +21,7 @@ public class PackageSanityTest extends AbstractPackageSanityTests {
     ignoreClasses(
         c ->
             c.equals(InterpolatingProverDelegate.class)
-                || c.equals(OptimizationProverDelegate.class));
+                || c.equals(OptimizationProverDelegate.class)
+                || c.equals(BasicProverWithAssumptionsWrapper.class));
   }
 }
